organize task based functions into their own file
