+++ pkg/skaffold/generate_pipeline/tasks_test.go
@@ -1,0 +1,111 @@
+/*
+Copyright 2019 The Skaffold Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package generatepipeline
+
+import (
+	"testing"
+
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
+	"github.com/GoogleContainerTools/skaffold/testutil"
+)
+
+func TestGenerateBuildTask(t *testing.T) {
+	var tests = []struct {
+		description string
+		buildConfig latest.BuildConfig
+		shouldErr   bool
+	}{
+		{
+			description: "successfully generate build task",
+			buildConfig: latest.BuildConfig{
+				Artifacts: []*latest.Artifact{
+					{
+						ImageName: "testArtifact",
+					},
+				},
+			},
+			shouldErr: false,
+		},
+		{
+			description: "fail generating build task",
+			buildConfig: latest.BuildConfig{
+				Artifacts: []*latest.Artifact{},
+			},
+			shouldErr: true,
+		},
+	}
+
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			configFile := &ConfigFile{
+				Name: "test",
+				Profile: &latest.Profile{
+					Pipeline: latest.Pipeline{
+						Build: test.buildConfig,
+					},
+				},
+			}
+			_, err := generateBuildTask(configFile)
+			t.CheckError(test.shouldErr, err)
+		})
+	}
+}
+
+func TestGenerateDeployTask(t *testing.T) {
+	var tests = []struct {
+		description  string
+		deployConfig latest.DeployConfig
+		shouldErr    bool
+	}{
+		{
+			description: "successfully generate deploy task",
+			deployConfig: latest.DeployConfig{
+				DeployType: latest.DeployType{
+					HelmDeploy: &latest.HelmDeploy{},
+				},
+			},
+			shouldErr: false,
+		},
+		{
+			description: "fail generating deploy task",
+			deployConfig: latest.DeployConfig{
+				DeployType: latest.DeployType{
+					HelmDeploy:      nil,
+					KubectlDeploy:   nil,
+					KustomizeDeploy: nil,
+				},
+			},
+			shouldErr: true,
+		},
+	}
+
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			configFile := &ConfigFile{
+				Name: "test",
+				Config: &latest.SkaffoldConfig{
+					Pipeline: latest.Pipeline{
+						Deploy: test.deployConfig,
+					},
+				},
+			}
+
+			_, err := generateDeployTask(configFile)
+			t.CheckError(test.shouldErr, err)
+		})
+	}
+}
